Moved TestRoles to the testFxitures
