+++ common/src/testFixtures/java/com/axelixlabs/axelix/common/testfixtures/TestRoles.java
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package com.axelixlabs.axelix.sbs.spring.core.utils;
+package com.axelixlabs.axelix.common.testfixtures;
 
 import java.util.Set;
